% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengToolkit.R
\name{add_function}
\alias{add_function}
\title{add funtions that you want to store them in a function.R file}
\usage{
add_function(..., source_fun = F)
}
\arguments{
\item{virable}{param my_functions input any functions that you want to}
}
\value{
information about what did the function do
}
\description{
today,I create my fourth function,a very useful function.
}
\details{
#' This is some description of this function.
#' @title update this package itself
#'
#' @description update this package from github
#'
#'
#'
#' @return nothing but love
#' @keywords set_mirror
#' @export
#' @examples
#' update_myself()
}
\section{update_myself <- function() {}{

}

\section{message("upgrading this package from github mugpeng/pengToolkit")}{

}

\section{if (grepl("pengToolkit", paste((.packages()), collapse = ", ")) == T) {}{

}

\section{detach("package:pengToolkit")}{

}

\section{}}{


you can use this function to receive any funtions that you want to store them in a function.R file
and source them automatically into your environment.
}

\section{devtools::install_github("mugpeng/pengToolkit")}{

}

\section{library(pengToolkit)}{

}

\examples{
add_function("test001", "test002")
add_function(source_fun = T)
}
\keyword{add_function}
