% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengToolkit.R
\name{boost_install_packages}
\alias{boost_install_packages}
\title{receive a vector of packages and install them from cran or bioconductor}
\usage{
boost_install_packages(my_packages = my_packages)
}
\arguments{
\item{my_packages}{input a vector of your packages}
}
\value{
information about what did the function do
}
\description{
today,I create my third function,a very usrful function.
}
\details{
This is some description of this function.

you can use this function to download a batch of uninstalled packages from CRAN or bioconductor with judment if exists.
}
\examples{
boost_install_packages(c("devtools", "roxygen2", "testthat"))
boost_install_packages(my_packages = c("devtools", "roxygen2", "testthat"))
}
\keyword{boost_install_packages}
