% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pengToolkit}
\alias{pengToolkit}
\title{pengToolkit: A package for store my useful functions.}
\description{
The pengToolkit package provides three cool functions:
boost_install_packages, set_mirror and make_r_template.
}
\references{
Full documentation and demos: \url{https://github.com/mugpeng/pengToolkit};
}
