% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengToolkit.R
\name{set_mirror}
\alias{set_mirror}
\title{set_mirror}
\usage{
set_mirror(loc = "China")
}
\arguments{
\item{loc}{input "recover" means drop the mirror}
}
\value{
nothing but love
}
\description{
set-up mirror in China
}
\details{
This is some description of this function.

you can use this function to change ur cran&&bioconductor mirrors if you are in China.
}
\examples{
set_mirror()
set_mirror("China")
}
\keyword{set_mirror}
