% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengToolkit.R
\name{update_myself}
\alias{update_myself}
\title{update this package itself}
\usage{
update_myself()
}
\value{
nothing but love
}
\description{
update this package from github
}
\details{
This is some description of this function.
}
\examples{
update_myself()
}
\keyword{set_mirror}
