% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengToolkit.R
\name{make_r_template}
\alias{make_r_template}
\title{A head comment template auto-generator}
\usage{
make_r_template(
  project = "Rescue the Princess",
  filename = "file.R",
  author = "Peng",
  dir = getwd()
)
}
\arguments{
\item{project}{input a character type or not}

\item{filename}{input a character type or not}

\item{author}{input a character type or not}

\item{dir}{input a character type or not}
}
\value{
file a new file with designed header
}
\description{
today,I create my first function,a very usrful function.
}
\details{
This is some description of this function.

you can use this function to create your header in the script.
}
\examples{
make_r_template()
make_r_template(project = "How to Fire My Boss.", filename = "01-preparation.R", author = "Tony", dir = getwd())
}
\keyword{make_r_template}
